+++ CHANGELOG.md
@@ -7,13 +7,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased]
+### Changed
 
+ - `serverline.init()` can accept an object `options`.
+
+### Added
+
+ - Add `colorMode` and `inspectOptions` properties to options in `serverline.init(options)`.
 
 
 ## [1.3.1] - 2019-08-20
 ### Fixed
 
  - Add `_debugModuleSupport(require('debug'))` function for debug module #6
+
 
 ## [1.3.0] - 2019-08-20
 ### Changed
@@ -23,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
  - Add `_debugModuleSupport(require('debug'))` function for debug module #5
+
 
 ## [1.2.0] - 2019-08-20
 ### Changed